Fix ErrorGroups

+++ modules/core/src/main/java/org/apache/ignite/lang/ErrorGroups.java
@@ -73,7 +73,7 @@ public class ErrorGroups {
         public static final int TABLE_DEFINITION_ERR = TABLE_ERR_GROUP.registerErrorCode((short) 6);
 
         /** Schema version mismatch. */
-        public static final int SCHEMA_VERSION_MISMATCH_ERR = TABLE_ERR_GROUP.registerErrorCode(7);
+        public static final int SCHEMA_VERSION_MISMATCH_ERR = TABLE_ERR_GROUP.registerErrorCode((short) 7);
     }
 
     /** Client error group. */
